Added gcc compiler flags and fixed compilation issues.

+++ CreateCMakeBuildDirectory.py
@@ -6,15 +6,30 @@ def main():
 
     configuration = "Debug"
 
-    if len(sys.argv) > 0:
-        if str(sys.argv)[0] == "Release":
+    if len(sys.argv) > 1:
+
+        if sys.argv[1] == "Release":
+
             configuration = "Release"
 
     directory = "Build" + configuration
+
     if path.exists(directory):
-        os.system('rm -rf {0}'.format(directory))
+
+        userInput = input("The {0} directory already exists. Do you want to delete it or not, please? (y/n)\n".format(directory))
+
+        if userInput.lower() == "y" or userInput.lower() == "yes":
+
+            os.system('rm -rf {0}'.format(directory))
+
+        else:
+
+            sys.exit("The {0} directory will remain intact.\n".format(directory))
+
     os.system('mkdir {0}'.format(directory))
+
     os.chdir('{0}'.format(directory))
+
     os.system('cmake ../ -DCMAKE_BUILD_TYPE={0}'.format(configuration))
 
 if __name__== "__main__":
